Refactor project package structure.
Brings query up one level from plugin to neuralsearch.
