+++ src/main/java/org/opensearch/neuralsearch/query/NeuralQueryBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.neuralsearch.plugin.query;
+package org.opensearch.neuralsearch.query;
 
 import static org.opensearch.neuralsearch.common.VectorUtil.vectorAsListToArray;
